various: flatten logged json, add skip-verify option
The intended use case for this app is running as a Kubernetes workload
with Graylog consuming its output.

In this scenario the nested JSON structure is not ideal at all, as
Graylog does not directly support working with arbitrarily nested
structures in log messages. Flattening the payload structure makes
alerting and dashboards much easier to achieve.

Additionally, emitting the logs in a format that is immediately useful
in Graylog avoids overloading Graylog with post-processing and
eliminates the need to configure any extractors/grok patterns.

Also added an option for skipping message digest verification. Don't
use this in production!